+++ sprite.cpp
@@ -3,7 +3,7 @@
 #include "globals.h"
 
 
-Sprite::Sprite(Graphic &graphic, int originX, int originY, int width, int height, float posX, float posY) : 
+Sprite::Sprite(Graphic &graphic, int originX, int originY, int width, int height, double posX, double posY) : 
 posX(posX), posY(posY) {
 	SDL_Surface * surface = graphic.getSurface("MyChar.png", false);
 	if(surface == nullptr){
@@ -22,10 +22,10 @@ posX(posX), posY(posY) {
 Sprite::~Sprite(){}
 	
 void Sprite::draw(Graphic &graphic){
-	SDL_Rect destination = {this->posX, this->posY, 
+	SDL_Rect destination = {(int)this->posX, (int)this->posY, 
 		this->source.w * globals::SPRITE_SCALER, this->source.h * globals::SPRITE_SCALER};
 	graphic.blitSurface(this->texture, &this->source, &destination);
 }
 
-void Sprite::update(float elapsedTime){}
+void Sprite::update(double elapsedTime){}
